#37287 #37353 (2.2.0.89) ALIGNMENT 型を定義し、型の置換を継続中。 / Define ALIGNMENT, ongoing type replacement.

+++ src/wizard1.c
@@ -333,7 +333,7 @@ static void kind_info(char *buf, char *dam, char *wgt, char *chance, DEPTH *lev,
 
 
 	/* Weight */
-	sprintf(wgt, "%3d.%d", q_ptr->weight / 10, q_ptr->weight % 10);
+	sprintf(wgt, "%3d.%d", (int)(q_ptr->weight / 10), (int)(q_ptr->weight % 10));
 }
